fix: Long 타입변경

+++ src/main/java/com/sparta/no1delivery/domain/user/application/UserService.java
@@ -13,7 +13,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigInteger;
 import java.util.UUID;
 
 @Service
@@ -45,14 +44,14 @@ public class UserService {
 
     // user 조회
     @Transactional(readOnly = true)
-    public User getUser(BigInteger userId) {
+    public User getUser(Long userId) {
 
         return userRepository.findById(userId)
                 .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
     }
 
     //닉네임 변경
-    public void changeNickname(BigInteger userId, String nickname) {
+    public void changeNickname(Long userId, String nickname) {
 
         User user = getUser(userId);
 
@@ -60,7 +59,7 @@ public class UserService {
     }
 
     //비밀번호 변경
-    public void changePassword(BigInteger userId, String password) {
+    public void changePassword(Long userId, String password) {
 
         User user = getUser(userId);
         String encodedPassword = passwordEncoder.encode(password);
@@ -69,7 +68,7 @@ public class UserService {
     }
 
     //회원 탈퇴
-    public void withdraw(BigInteger userId){
+    public void withdraw(Long userId){
         User user = getUser(userId);
 
         //회원 탈퇴 메소드
@@ -77,7 +76,7 @@ public class UserService {
 
     // 주소 추가
     public void addAddress(
-            BigInteger userId,
+            Long userId,
             String address,
             String detailAddress,
             Boolean isDefault
@@ -98,7 +97,7 @@ public class UserService {
 
     //주소 수정
     public void updateAddress(
-            BigInteger userId,
+            Long userId,
             UUID addressId,
             String address,
             String detailAddress
@@ -119,7 +118,7 @@ public class UserService {
 
     //주소 삭제
     public void deleteAddress(
-            BigInteger userId,
+            Long userId,
             UUID addressId
     ) {
 
@@ -137,7 +136,7 @@ public class UserService {
 
     //기본 배송지 변경
     public void changeDefaultAddress(
-            BigInteger userId,
+            long userId,
             UUID addressId
     ) {
 
@@ -154,7 +153,7 @@ public class UserService {
     }
 
     //사장 권한 요청
-    public void requestOwnerRole(BigInteger userId, String businessNumber) {
+    public void requestOwnerRole(Long userId, String businessNumber) {
 
         User user = getUser(userId);
 
@@ -162,7 +161,7 @@ public class UserService {
     }
 
     //사장 권한 승인
-    public void approveOwnerRole(BigInteger userId) {
+    public void approveOwnerRole(Long userId) {
 
         User user = getUser(userId);
 
@@ -170,7 +169,7 @@ public class UserService {
     }
 
     // 사장 권한 거절
-    public void rejectOwnerRole(BigInteger userId) {
+    public void rejectOwnerRole(Long userId) {
 
         User user = getUser(userId);
 
@@ -179,7 +178,7 @@ public class UserService {
 
 
     //사장 → 손님 권한 다운그레이드
-    public void downgradeToCustomer(BigInteger userId) {
+    public void downgradeToCustomer(Long userId) {
 
         User user = getUser(userId);
